Little reorganization of EventProjection, getting ready for extra tables
