+++ src/Marten.AsyncDaemon.Testing/build_aggregate_projection.cs
@@ -121,6 +121,35 @@ namespace Marten.AsyncDaemon.Testing
         }
 
         [Fact]
+        public async Task rebuild_the_projection_clears_state()
+        {
+            NumberOfStreams = 1;
+
+            Logger.LogDebug("The expected number of events is {NumberOfEvents}", NumberOfEvents);
+
+            StoreOptions(x => x.Projections.Add(new TripAggregationWithCustomName(), ProjectionLifecycle.Async), true);
+
+            var agent = await StartDaemon();
+
+            await PublishSingleThreaded();
+
+            var waiter = agent.Tracker.WaitForShardState(new ShardState("Trip:All", NumberOfEvents), 30.Seconds());
+
+            await waiter;
+
+            // This should be gone after a rebuild
+            var trip = new Trip();
+            theSession.Store(trip);
+            await theSession.SaveChangesAsync();
+
+            await agent.RebuildProjection("Trip", CancellationToken.None);
+
+            using var query = theStore.QuerySession();
+            // Demonstrates that the Trip documents were deleted first
+            (await query.LoadAsync<Trip>(trip.Id)).ShouldBeNull();
+        }
+
+        [Fact]
         public async Task rebuild_the_projection_without_custom_name()
         {
             NumberOfStreams = 10;
